refactor: Condition과 Status 이름 변경
