+++ backend/src/main/java/com/zzang/chongdae/comment/service/CommentService.java
@@ -62,7 +62,7 @@ public class CommentService {
         OfferingEntity offering = offeringRepository.findById(offeringId)
                 .orElseThrow(() -> new MarketException(OfferingErrorCode.NOT_FOUND));
         validateIsJoined(member, offering);
-        if (offering.isStatusGrouping() && offering.toOfferingStatus().isAutoConfirmed()) {
+        if (offering.isStatusGrouping() && offering.toOfferingCondition().isAutoConfirmed()) {
             offering.moveStatus();
         }
         return new CommentRoomInfoResponse(offering, member);
