+++ backend/src/main/java/com/zzang/chongdae/offering/service/dto/OfferingAllResponseItem.java
@@ -1,8 +1,8 @@
 package com.zzang.chongdae.offering.service.dto;
 
-import com.zzang.chongdae.offering.domain.OfferingCondition;
-import com.zzang.chongdae.offering.domain.OfferingPrice;
 import com.zzang.chongdae.offering.domain.OfferingStatus;
+import com.zzang.chongdae.offering.domain.OfferingPrice;
+import com.zzang.chongdae.offering.domain.OfferingCondition;
 import com.zzang.chongdae.offering.repository.entity.OfferingEntity;
 
 public record OfferingAllResponseItem(Long id,
@@ -13,21 +13,21 @@ public record OfferingAllResponseItem(Long id,
                                       String thumbnailUrl,
                                       Integer dividedPrice,
                                       Integer originPrice,
-                                      OfferingCondition condition,
+                                      OfferingStatus status,
                                       Boolean isOpen) {
 
     public OfferingAllResponseItem(
-            OfferingEntity offering, OfferingPrice offeringPrice, OfferingStatus offeringStatus) {
+            OfferingEntity offering, OfferingPrice offeringPrice, OfferingCondition offeringCondition) {
         this(offering.getId(),
                 offering.getTitle(),
                 offering.getMeetingAddressDong(),
-                offeringStatus.getCurrentCount(),
+                offeringCondition.getCurrentCount(),
                 offering.getTotalCount(),
                 offering.getThumbnailUrl(),
                 offeringPrice.calculateDividedPrice(),
                 offeringPrice.getOriginPrice(),
-                offeringStatus.decideOfferingCondition(),
-                offeringStatus.isOpen()
+                offeringCondition.decideOfferingStatus(),
+                offeringCondition.isOpen()
         );
     }
 }
